requirements.txt
numpy
pillow
fastapi
ultralytics
typing
uvicorn